refactored benzinga provider directory
