+++ openbb_sdk/providers/benzinga/openbb_benzinga/provider.py
@@ -7,8 +7,8 @@
 # IMPORT INTERNAL
 from openbb_provider.abstract.provider import Provider, ProviderNameType
 
-from openbb_benzinga.global_news import BenzingaGlobalNewsFetcher
-from openbb_benzinga.stock_news import BenzingaStockNewsFetcher
+from openbb_benzinga.models.global_news import BenzingaGlobalNewsFetcher
+from openbb_benzinga.models.stock_news import BenzingaStockNewsFetcher
 
 # mypy: disable-error-code="list-item"
